tests/ports/psoc-edge: Simplified bitstream test cases.
