+++ tests/ports/psoc-edge/board_ext_hw/multi/bitstream.py.exp
@@ -1,3 +1,7 @@
+--- instance0 ---
 bitstream call ok: True
 bitstream rejects short timing: True
 bitstream tx ok: True
+--- instance1 ---
+bitstream rx ok: True
+data match: True
