--- tests/ports/psoc-edge/board_ext_hw/multi/bitstream_rx.py.exp
@@ -1,2 +1,0 @@
-bitstream rx ok: True
-data match: True
