+++ tests/ports/psoc-edge/test-plan.yml
@@ -44,16 +44,9 @@
 #           - pin
 
 - name: bitstream
-  type: multi_stub
+  type: multi
   test:
-    script: ports/psoc-edge/board_ext_hw/multi/bitstream_rx.py
-    post_stub_delay_ms: 1000
-    device:
-      - board: KIT_PSE84_AI
-        features:
-          - bitstream
-  stub:
-    script: ports/psoc-edge/board_ext_hw/multi/bitstream_tx.py
+    script: ports/psoc-edge/board_ext_hw/multi/bitstream.py
     device:
       - board: KIT_PSE84_AI
         features:
